Introduce TracePoint formatters and refactor trace data handling
