+++ lib/trace_viz/formatters/trace_point/base_formatter.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require "trace_viz/formatters/base_formatter"
+
+module TraceViz
+  module Formatters
+    module TracePoint
+      class BaseFormatter < TraceViz::Formatters::BaseFormatter
+      end
+    end
+  end
+end
